move PlaylistContentView into its own file
